Fix prototypes to strictly match Arduino API
See https://github.com/arduino/Arduino/issues/4525

+++ cores/arduino/delay.h
@@ -33,7 +33,7 @@ extern "C" {
  *
  * \return Number of milliseconds since the program started (uint32_t)
  */
-extern uint32_t millis( void ) ;
+extern unsigned long millis( void ) ;
 
 /**
  * \brief Returns the number of microseconds since the Arduino board began running the current program.
@@ -45,7 +45,7 @@ extern uint32_t millis( void ) ;
  *
  * \note There are 1,000 microseconds in a millisecond and 1,000,000 microseconds in a second.
  */
-extern uint32_t micros( void ) ;
+extern unsigned long micros( void ) ;
 
 /**
  * \brief Pauses the program for the amount of time (in miliseconds) specified as parameter.
@@ -53,15 +53,15 @@ extern uint32_t micros( void ) ;
  *
  * \param dwMs the number of milliseconds to pause (uint32_t)
  */
-extern void delay( uint32_t dwMs ) ;
+extern void delay( unsigned long dwMs ) ;
 
 /**
  * \brief Pauses the program for the amount of time (in microseconds) specified as parameter.
  *
  * \param dwUs the number of microseconds to pause (uint32_t)
  */
-static __inline__ void delayMicroseconds( uint32_t ) __attribute__((always_inline, unused)) ;
-static __inline__ void delayMicroseconds( uint32_t usec )
+static __inline__ void delayMicroseconds( unsigned int ) __attribute__((always_inline, unused)) ;
+static __inline__ void delayMicroseconds( unsigned int usec )
 {
   if ( usec == 0 )
   {
